create listbox && change combobox to use it
